requirements.txt
streamlit
langchain
langchain_ollama
langchain_community
streamlit-chat
pypdf
chromadb
python-dotenv
langchain-groq
langchain-google-genai
langchain-openai
